+++ ESFramework/Core/Additions/NSArray+ESAdditions.m
@@ -10,11 +10,6 @@
 
 @implementation NSArray (ESAdditions)
 
-- (BOOL)isEmpty
-{
-    return (0 == self.count);
-}
-
 - (id)objectPassingTest:(BOOL (^)(id, NSUInteger, BOOL *))predicate
 {
     return [self objectWithOptions:0 passingTest:predicate];